moving modules _manager into dedicated sub module
