+++ packages/pymodaq/src/pymodaq/extensions/daq_logger/h5logging.py
@@ -12,7 +12,7 @@ from pymodaq_utils.logger import set_logger, get_module_name
 from pymodaq_gui.h5modules.saving import H5Saver
 from pymodaq_data.data import DataToExport
 
-from pymodaq.utils.managers.modules_manager import ModulesManager
+from pymodaq.utils.managers.modules.modules_manager import ModulesManager
 from pymodaq.utils.h5modules import module_saving
 from .abstract import AbstractLogger
