+++ packages/pymodaq/src/pymodaq/utils/managers/__init__.py
@@ -1,0 +1,1 @@
+from .modules import modules_manager
