+++ packages/pymodaq/src/pymodaq/utils/managers/modules/__init__.py
@@ -1,0 +1,2 @@
+from .modules_manager import ModulesManager
+from .utils import ModuleType
