+++ packages/pymodaq/src/pymodaq/utils/managers/modules/utils.py
@@ -1,0 +1,8 @@
+from pymodaq_utils.enums import StrEnum
+
+
+class ModuleType(StrEnum):
+    Actuator = "actuator"
+    Detector = "detector"
+    Other = 'other'
+    NONE = 'None'
