+++ packages/pymodaq/src/pymodaq/utils/managers/preset/utils.py
@@ -10,7 +10,7 @@ from pymodaq_gui.parameter.utils import get_param_dict_from_name
 from pymodaq.control_modules.instruments import DET_TYPES, ACTUATOR_TYPES, ACTUATOR_NAMES
 from pymodaq.control_modules.daq_move_ui.factory import ActuatorUIFactory
 from pymodaq.control_modules.utils import create_controller_param
-from pymodaq.utils.managers.modules_manager import ModuleType
+from pymodaq.utils.managers.modules.utils import ModuleType
 
 config = Config()
 logger = set_logger(get_module_name(__file__))
